Support for port allocation request for NIF on Network.
For every VM network interface backed by a vsphere distributed port group network,
vCenter server as part of VM lifeycle operations allocates a Port (and corresponding ConnectionID)
on the network. This information is used on the host to attach the VM network interface to the network.
There are cases when we want to by-pass vCenter server for VM lifeycle operations. In such cases
we need to explicitly request vCenter server to allocate a Port for the network interface on
the vsphere distributed port group network. This change enables such a request
to be made via NetworkInterfaceSpec.

Testing Done:
make generate