Remove namespace from NetworkReference
Networks are provided to namespaces through VI admin configuration.
Since only authorized networks are provided in the user's namespace, it
does not make sense to allow users to select other namespaces that they
may be unauthorized to read. Instead, the VI admin would be expected to
make the network available to that namespace.